requirements.txt
# local package
-e .

# external requirements
click~=8.1.2
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

pandas~=1.4.2
setuptools~=60.10.0
numpy~=1.22.3
src~=0.1.0
Hydra~=2.5
sklearn~=0.0
scikit-learn~=1.0.2
mlflow~=1.25.1
omegaconf~=2.1.2